Use binary data as HMAC key
The application key stored in base64 in the database is now decoded
before being used as key in HMAC. This fixes signature validation
problems with Auth/Yubico.php, used - for example - in simpleid-yubikey.

Previously, the ASCII representation of the base64 encoded key was used
as HMAC key, which generated different signatures. This change is not
backwards compatible: it will break any clients that have so far
successfully validated the signatures.

Since the date and time of application creation was used as seed for the
generated key, but a string is not easily convertible into raw bytes
(and is potentially low-entropy), switch to 256 (cryptographically
secure) random bytes as key initialization.

See neverpanic/simpleid-yubikey#1

Signed-off-by: Clemens Lang <neverpanic@gmail.com>